Corected tests
