+++ tests/integration/Mvc/Model/MetaData/Libmemcached/WriteCest.php
@@ -12,13 +12,22 @@ declare(strict_types=1);
 
 namespace Phalcon\Test\Integration\Mvc\Model\MetaData\Libmemcached;
 
+use function getOptionsLibmemcached;
 use IntegrationTester;
+use Phalcon\Cache\AdapterFactory;
+use Phalcon\Mvc\Model\Exception;
+use Phalcon\Mvc\Model\MetaData\Libmemcached;
+use Phalcon\Storage\SerializerFactory;
+use Phalcon\Test\Fixtures\Traits\DiTrait;
+use Phalcon\Test\Models\Robots;
 
 /**
  * Class WriteCest
  */
 class WriteCest
 {
+    use DiTrait;
+
     /**
      * Tests Phalcon\Mvc\Model\MetaData\Libmemcached :: write()
      *
@@ -29,5 +38,50 @@ class WriteCest
     {
         $I->wantToTest('Mvc\Model\MetaData\Libmemcached - write()');
         $I->skipTest('Need implementation');
+    }
+
+    /**
+     * Tests Phalcon\Mvc\Model\MetaData\Libmemcached :: write() - exception
+     *
+     * @author Phalcon Team <team@phalconphp.com>
+     * @since  2018-11-13
+     */
+    public function mvcModelMetadataLibmemcachedWriteException(IntegrationTester $I)
+    {
+        $I->wantToTest('Mvc\Model\MetaData\Libmemcached - write() - exception');
+        $this->setNewFactoryDefault();
+        $this->setDiMysql();
+        $this->container->set(
+            'modelsMetadata',
+            function () {
+                $options                       = getOptionsLibmemcached();
+                $options['servers'][0]['port'] = 4000;
+                $serializer                    = new SerializerFactory();
+                $factory                       = new AdapterFactory($serializer);
+                return new Libmemcached($factory, $options);
+            }
+        );
+        $I->expectThrowable(
+            new Exception('Failed to store metaData to the cache adapter'),
+            function () use ($I) {
+                Robots::setup(
+                    [
+                        'exceptionOnFailedMetaDataSave' => true,
+                    ]
+                );
+                Robots::findFirst(1);
+            }
+        );
+        $I->expectThrowable(
+            new \Exception('Failed to store metaData to the cache adapter', 1024),
+            function () use ($I) {
+                Robots::setup(
+                    [
+                        'exceptionOnFailedMetaDataSave' => false,
+                    ]
+                );
+                Robots::findFirst(1);
+            }
+        );
     }
 }
